pyproject.toml
[tool.poetry]
name = "het-veldje"
version = "0.1.0"
description = "FastAPI app voor info omtrent het veldje in kanaleneiland"
authors = ["fliepeltje <donatas@rasiukevicius.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
jinja2 = "^3.0.1"
aiofiles = "^0.7.0"
uvicorn = {extras = ["standard"], version = "^0.14.0"}
toml = "^0.10.2"
httpx = "^0.18.2"
pytz = "^2021.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"
pdbpp = "^0.10.3"
bpython = "^0.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
